added health bar

+++ src/Application/include/widgets/helpers.hpp
@@ -17,7 +17,7 @@ namespace helper {
 
 // path relative to the data directory
 auto getTexture(const std::string &simplePath) -> std::uint32_t;
-auto from1080p(int x, int y) noexcept -> ImVec2;
+auto from1080p(float x, float y) noexcept -> ImVec2;
 
 auto frac2pixel(ImVec2 fraction) noexcept -> ImVec2;
